+++ lib/action_view/component/test_helpers.rb
@@ -3,8 +3,8 @@
 module ActionView
   module Component
     module TestHelpers
-      def render_component(component, &block)
-        Nokogiri::HTML(component.render_in(ApplicationController.new.view_context, &block))
+      def render_component(options = {}, &block)
+        Nokogiri::HTML(options[:component].new(options[:locals]).render_in(ApplicationController.new.view_context, &block))
       end
     end
   end