+++ test/action_view/component_test.rb
@@ -6,14 +6,14 @@ class ActionView::ComponentTest < Minitest::Test
   include ActionView::Component::TestHelpers
 
   def test_render_component
-    result = render_component(MyComponent.new)
+    result = render_component(component: MyComponent)
 
     assert_equal trim_result(result.css("div").first.to_html), "<div>hello,world!</div>"
   end
 
   def test_raises_error_when_sidecar_template_is_missing
     exception = assert_raises NotImplementedError do
-      render_component(MissingTemplateComponent.new)
+      render_component(component: MissingTemplateComponent)
     end
 
     assert_includes exception.message, "Could not find a template file for MissingTemplateComponent"
@@ -21,7 +21,7 @@ class ActionView::ComponentTest < Minitest::Test
 
   def test_raises_error_when_more_then_one_sidecar_template_is_present
     error = assert_raises StandardError do
-      render_component(TooManySidecarFilesComponent.new)
+      render_component(component: TooManySidecarFilesComponent)
     end
 
     assert_includes error.message, "More than one template found for TooManySidecarFilesComponent."
@@ -29,7 +29,7 @@ class ActionView::ComponentTest < Minitest::Test
 
   def test_raises_error_when_initializer_is_not_defined
     exception = assert_raises NotImplementedError do
-      render_component(MissingInitializerComponent.new)
+      render_component(component: MissingInitializerComponent)
     end
 
     assert_includes exception.message, "must implement #initialize"
@@ -37,14 +37,14 @@ class ActionView::ComponentTest < Minitest::Test
 
   def test_checks_validations
     exception = assert_raises ActiveModel::ValidationError do
-      render_component(WrapperComponent.new)
+      render_component(component: WrapperComponent)
     end
 
     assert_includes exception.message, "Content can't be blank"
   end
 
   def test_renders_content_from_block
-    result = render_component(WrapperComponent.new) do
+    result = render_component(component: WrapperComponent) do
       "content"
     end
 
@@ -52,28 +52,28 @@ class ActionView::ComponentTest < Minitest::Test
   end
 
   def test_renders_slim_template
-    result = render_component(SlimComponent.new(message: "bar")) { "foo" }
+    result = render_component(component: SlimComponent, locals: { message: "bar" }) { "foo" }
 
     assert_includes result.text, "foo"
     assert_includes result.text, "bar"
   end
 
   def test_renders_haml_template
-    result = render_component(HamlComponent.new(message: "bar")) { "foo" }
+    result = render_component(component: HamlComponent, locals: { message: "bar" }) { "foo" }
 
     assert_includes result.text, "foo"
     assert_includes result.text, "bar"
   end
 
   def test_renders_erb_template
-    result = render_component(ErbComponent.new(message: "bar")) { "foo" }
+    result = render_component(component: ErbComponent, locals: { message: "bar" }) { "foo" }
 
     assert_includes result.text, "foo"
     assert_includes result.text, "bar"
   end
 
   def test_renders_route_helper
-    result = render_component(RouteComponent.new)
+    result = render_component(component: RouteComponent)
 
     assert_includes result.text, "/"
   end
@@ -81,25 +81,25 @@ class ActionView::ComponentTest < Minitest::Test
   def test_template_changes_are_not_reflected_in_production
     ActionView::Base.cache_template_loading = true
 
-    assert_equal "<div>hello,world!</div>", render_component(MyComponent.new).css("div").first.to_html
+    assert_equal "<div>hello,world!</div>", render_component(component: MyComponent).css("div").first.to_html
 
     modify_file "app/components/my_component.html.erb", "<div>Goodbye world!</div>" do
-      assert_equal  "<div>hello,world!</div>", render_component(MyComponent.new).css("div").first.to_html
+      assert_equal  "<div>hello,world!</div>", render_component(component: MyComponent).css("div").first.to_html
     end
 
-    assert_equal "<div>hello,world!</div>", render_component(MyComponent.new).css("div").first.to_html
+    assert_equal "<div>hello,world!</div>", render_component(component: MyComponent).css("div").first.to_html
   end
 
   def test_template_changes_are_reflected_outside_production
     ActionView::Base.cache_template_loading = false
 
-    assert_equal "<div>hello,world!</div>", render_component(MyComponent.new).css("div").first.to_html
+    assert_equal "<div>hello,world!</div>", render_component(component: MyComponent).css("div").first.to_html
 
     modify_file "app/components/my_component.html.erb", "<div>Goodbye world!</div>" do
-      assert_equal "<div>Goodbye world!</div>", render_component(MyComponent.new).css("div").first.to_html
+      assert_equal "<div>Goodbye world!</div>", render_component(component: MyComponent).css("div").first.to_html
     end
 
-    assert_equal "<div>hello,world!</div>", render_component(MyComponent.new).css("div").first.to_html
+    assert_equal "<div>hello,world!</div>", render_component(component: MyComponent).css("div").first.to_html
   end
 
   private